Properly clean up constraints when node leaves graph
Before this change, some node that carried constraints would remain
registered as interested in pending dependencies. This could cause
invalid deferred selection.
With this set of changes, such state is cleaned up when a node that
carries constraints no longer participates in the graph, either because
it is not transitive or because it is no longer selected.

Issue #13329